adding error pages

+++ src/main/java/com/filemanager/repositories/LogRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface LogRepository  extends JpaRepository<Log, Integer> , PagingAndSortingRepository<Log, Integer> {
+public interface LogRepository  extends JpaRepository<Log, Long> , PagingAndSortingRepository<Log, Long> {
     Page<Log> findAllByOrderByDateDesc(Pageable pageable);
     int countAllByMessageContaining(String message);
 }
